Support query param in upstream value
